fix: identical names for separate `azurerm_monitor_diagnostic_setting` resources (#178)
Currently, both the main module and the `database` submodule create a
resource named `azurerm_monitor_diagnostic_setting.this`.

This looks confusing on the Terraform Registry page for this module, as
it looks like this module creates the same resource twice.

Rename the `azurerm_monitor_diagnostic_setting.this` resource in the
main module to `azurerm_monitor_diagnostic_setting.server` to reflect
that this diagnostic setting is scoped to the SQL server.

Rename the `azurerm_monitor_diagnostic_setting.this` resource in the
`database` submodule to `azurerm_monitor_diagnostic_setting.database` to
reflect that this diagnostic setting is scoped to the SQL database.